Fixed Connection failing for non-ssl connections
Fixes #34

+++ include/libCom/network/Connection.h
@@ -131,7 +131,7 @@ public:
      * @param size Maximum bytes to read
 	 * @return >0: Actual bytes read. 0: (clean) shutdown. <0: error occurred
 	 */
-    int32_t readMax(Buffer &buffer, const uint32_t size);
+    ssize_t readMax(Buffer &buffer, const uint32_t size);
 
     /**
      * Read exactly size bytes from remote into the buffer. Blocks while waiting
